move stuff to lib/
