+++ lib/transform-options.js
