+++ lib/worker.js
